requirements.txt
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.14.0
Flask==3.0.3
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
sympy==1.12
tbb==2021.12.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.2
transformers==4.40.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
